Add support for Pico v2 and uEVK
Added support for version 2 of the Pico board. Imported the i.MX8 media
driver changes and the AP1302 driver from the 5.10.52 branch.

Added battery detection to the BQ2589x driver. On start up, the driver
samples the state of the charge ten times in quick succession. If the
state changes more than once, the battery is determined to be missing
and charging is disabled. Also added a sysfs attribute that allows for
direct control of the charge state.