Add support for Pico v3 (#12)
Added support for the Pico v3 board and its peripherals.

Added new device drivers for the MMC5633 magnetometer and SHT4x
temperature and humidity sensor.

Added basic support for the STAT output pin to the BQ2589x driver.

Added support for the SPI interface to the AP1302 driver; this has been
used to reduce the firmware loading time to 0.5 seconds.

Co-authored-by: WilliamBagshaw-imd <110393051+WilliamBagshaw-imd@users.noreply.github.com>