Support for IMDT DSI panel  (#102)
This commit adds support for IMDT DSI pannel IM-MSC-0006-01.
Changes:
1. In the device tree added a panel node under mipi_dsi node which uses the simple panel driver
   The back light is enabled via GPIO01_01 and is fixed as a gpio hog.
   Panel reset is via GPIO1_07
   Power is from the peripheral power domain which is enabled.
   For now, leaving HDMI support in the device tree.  It is untested which panel will be used if both are connected.

2. Added support for the panel in the panel-simple.c driver. Properties are based on the panel specification.
Width/height parameters in the drm_display_info struct under the panel_desc_dsi to use mm and not pixels. This is for compatibility with applications that may use these values (e.g. Weston) via the DRM APIs to calculate size in pixels.